Add files via upload
This is a basic example of how to use several different libraries for classification and ensembling, mostly with sklearn. Example assumes the use of a retail customer dataset where the user is targeting a binary variable for sale/no sale. I highly recommend running this in Jupyter Notebook or Lab, or potentially in Colab.

Some code was borrowed, with love, from the notebooks of the Kaggle community.